fix(regression): promote an integer design to float before the covariance
`np.asarray(X)` kept an integer dtype, so `X.T @ X` wrapped for designs with
large integer entries. `pinv` promotes, so the coefficients were right while
the standard errors, t-statistics and p-values were silently wrong — a
t of 0.71 where the same design in float gives 3.87.

Sweep: G-04

Changelog: `regress` promotes integer `X` and `Y` to float, so a large integer
design no longer overflows into meaningless standard errors and p-values.

+++ nltools/algorithms/regression.py
@@ -40,8 +40,11 @@ def regress(X, Y, *, stats: str = "full", tail: int | str = 2):
         raise ValueError("stats must be one of 'full', 'betas', 'tstats'")
     tail_internal = _validate_tail_parameter(tail)
 
-    X = np.asarray(X)
-    Y = np.asarray(Y)
+    # Promote to float before the covariance: an integer design overflows in
+    # `X.T @ X` (a column of 1e5 wraps in int32), which leaves `b` correct —
+    # `pinv` promotes — while se, t and p come out garbage with no warning.
+    X = np.asarray(X, dtype=float)
+    Y = np.asarray(Y, dtype=float)
     y_was_1d = Y.ndim == 1
     if y_was_1d:
         Y = Y[:, np.newaxis]
